Refactored UI
1. Rewrote UI in an FXML file instead of in the Java code itself, thus
separating the UI from the application logic.
2. Refactored the GUI class to be a singleton, so that all other classes
can access the Alert functionality to display alerts on the GUI.
3. Added an ErrorLogger class for logging non-critical errors on the
TextArea on the GUI.
