+++ src/ErrorLogger.java
@@ -1,0 +1,10 @@
+/**
+ * Created by harsh on 2017-04-08.
+ */
+public class ErrorLogger {
+
+    public static void log(String error) {
+//        GUI.getGui()
+        System.out.println("dummy");
+    }
+}
