+++ src/GUI.fxml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.geometry.Insets?>
+
+<?import javafx.scene.text.Text?>
+<GridPane alignment="CENTER" hgap="10.0" vgap="10.0" fx:id="grid"
+          xmlns:fx="http://javafx.com/fxml"
+          fx:controller="GUI">
+    <padding>
+        <Insets bottom="25.0" left="25.0" right="25.0" top="25.0" />
+    </padding>
+    <Text style="-fx-font: NORMAL 18 Tahoma;" text="Select a location to copy images"
+          GridPane.columnIndex="0" GridPane.rowIndex="0">
+    </Text>
+    <Button fx:id="browseButton" text="Browse..." GridPane.columnIndex="0" GridPane.rowIndex="1">
+    </Button>
+    <Label fx:id="path" text="" GridPane.columnIndex="1" GridPane.rowIndex="1">
+    </Label>
+    <Button fx:id="okButton" text="Go" GridPane.columnIndex="0" GridPane.rowIndex="3" visible="false">
+    </Button>
+    <!--<TextArea fx:id="logArea" style="-fx-font: NORMAL 18 Tahoma;" visible="false">-->
+    <!--</TextArea>-->
+
+</GridPane>
